% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcica_bss_dwst.R
\name{gcica_bss_dwst}
\alias{gcica_bss_dwst}
\title{Implement different W same T group-colorICA algorithm.}
\usage{
gcica_bss_dwst(Xc)
}
\arguments{
\item{Xc}{G by NG cell-array contains observed         [cell-array]
data matrices. G is the number of groups and NG is the number of subjects in each group.
Each component of the cell-array is a matrix with dimension M by T
where M is the number of mixtures and T is the number of time points.
Each data matrix has been pre-whitenned and centered.}
}
\value{
W: M by M unmixitng matrix
}
\description{
Implement different W same T group-colorICA algorithm.
}
\examples{
x=rnorm(10)
}
